+++ packages/components/src/Tree/stories/ColorfulTree.story.tsx
@@ -1,0 +1,59 @@
+/*
+
+ MIT License
+
+ Copyright (c) 2020 Looker Data Sciences, Inc.
+
+ Permission is hereby granted, free of charge, to any person obtaining a copy
+ of this software and associated documentation files (the "Software"), to deal
+ in the Software without restriction, including without limitation the rights
+ to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ copies of the Software, and to permit persons to whom the Software is
+ furnished to do so, subject to the following conditions:
+
+ The above copyright notice and this permission notice shall be included in all
+ copies or substantial portions of the Software.
+
+ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ SOFTWARE.
+
+ */
+
+import React from 'react'
+import { Tree, TreeGroup, TreeItem } from '../..'
+
+export const ColorfulTree = () => (
+  <Tree defaultOpen label="Colorful Tree">
+    <TreeGroup color="red" label="Red TreeGroup">
+      <TreeItem icon="Calendar">
+        Red TreeItem colored by parent TreeGroup's color prop
+      </TreeItem>
+      <TreeItem icon="Calendar">
+        Red TreeItem colored by parent TreeGroup's color prop
+      </TreeItem>
+      <TreeItem color="green" icon="Calendar">
+        Green TreeItem overriding parent color
+      </TreeItem>
+    </TreeGroup>
+    <TreeGroup
+      color="orange"
+      labelColor="blue"
+      label="Orange TreeGroup with 'blue' on labelColor prop"
+    >
+      <TreeItem icon="Calendar">
+        Orange TreeItem colored by parent TreeGroup's color prop
+      </TreeItem>
+      <TreeItem icon="Calendar">
+        Orange TreeItem colored by parent TreeGroup's color prop
+      </TreeItem>
+      <TreeItem color="purple" icon="Calendar">
+        Purple TreeItem overriding parent color
+      </TreeItem>
+    </TreeGroup>
+  </Tree>
+)